feat: setup gRPC infrastructure
This adds all of the base functionality needed to
develop gRPC based services. The gRPC service uses
an in-memory listener to avoid exposing a gRPC
service externally.

The gRPC RESTful proxy will be called by the echo
router when a 404 is encountered. One call out is
that this results in the 404 response being
different from current behavior.

The `third_party` folder will likely be replaced
with Buf in the future, just copied the protos for
now.