Demonstrate gRPC Rest Proxy Integration (#58)
* feat: setup gRPC infrastructure

This adds all of the base functionality needed to
develop gRPC based services. The gRPC service uses
an in-memory listener to avoid exposing a gRPC
service externally.

The gRPC RESTful proxy will be called by the echo
router when a 404 is encountered. One call out is
that this results in the 404 response being
different from current behavior.

The `third_party` folder will likely be replaced
with Buf in the future, just copied the protos for
now.

* feat: define projects API

* feat: generate projects API code and documentation

* feat: add projects API implementation

* chore: rename package

* chore: rename package

* bugfix: fix project service registration

The project service needs to be registered before
the gRPC service is started.

* feat: add required permissions annotation

Also updates the http bindings to use
`resourcemanager` as the path prefix to namespace
the service endpoints.

* Remove grpc usage

* Remove grpc usage

* Remove grpc usage

* Remove oapiv2 spec in favor of oapiv3

* Remove oapiv2 protoc gen in favor of oapiv3

* Adjust Taskfiles

* Add default task

* Add default task

* Add gnostic

* Remove iterative proto files for now, we will add them back in later

* Add datum-clients directory for later use.

* Remove for now

* No geodetic

* go mod tidy

* Remove workspace since we will generate directly into the service http server handlers folder.

---------

Co-authored-by: Aus Lacroix <aus@littlebearlabs.io>