chore: fix protobuf generate command
The proto-gen-openapi command wasn't being
installed in the grpc installation task so running
`api:generate-proto` was failing.

This also removes the old
`api:generate-http-api-docs` command and renames
`api:generate-gnostic` to `api:generate-openapi`
since that better describes the goal of the task.

This also adjusts the go module that's used for
the generated code to have a namespace specific to
the datum-os repo so it can be namespaced
differently than datum-cloud.